Change LinkedBinding to not be a Provider
We are going to start handling requests for Provider<T> soon and this will be confusing. A LinkedBinding<Provider<T>> would be a Provider<Provider<T>> then and I don't want to get caught up in accidentally interpreting the LinkedBinding itself as a Provider which it is not (providers should link lazily).